Extract PrivateInvoke class.
